Changing GPIO import to allow importing on non-RPi systems
Currently, dht11.py can only be imported as a package on a Raspberry Pi
system, even if the client does not use anything in the package that requires
a Raspberry Pi. This hinders unit testing and continuous integration systems
that might want to load the module, but not exercise any functionality that
requires an actual Raspberry Pi.

This change fixes the issue by changing the import of RPi.GPIO to only import
the RPi module and refer to GPIO by the fully-qualified RPi.GPIO.